change folders/components structure & add reactstrap-grid
