--- src/App.tsx
@@ -1,14 +1,0 @@
-import React from 'react';
-import logo from './logo.svg';
-import './App.css';
-import { Layout } from './layout/layout';
-
-const App = () => {
-  return (
-    <div className="App">
-      <Layout />
-    </div>
-  );
-};
-
-export default App;
