+++ src/components/user/user.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import './user.css';
+import './user.scss';
 
 export type UserType = {
   name: string;
@@ -9,9 +9,7 @@ export const User = ({ name }: UserType) => {
   return (
     <div className="user">
       <div className="user__status" />
-      <div className="user__name">{name}</div>
+      <div>{name}</div>
     </div>
   );
 };
-
-// export default User;
