+++ src/components/users_bar/users_bar.scss
@@ -1,14 +1,18 @@
+$header: rgb(0, 188, 212);
+$header_font_color: white;
+
 .users-bar {
     min-width: 20%;
     box-shadow: 1px 0 5px grey;
+    height: 100%;
 }
 
 .users-bar__header {
-    background-color: rgb(0, 188, 212);
+    background-color: $header;
     font-size: x-large;
     padding: 1rem 0 1rem 1rem;
     text-transform: capitalize;
-    color: white;
+    color: $header_font_color;
 }
 
 .users__user-list {
