--- src/layout/layout.css
@@ -1,10 +1,0 @@
-.layout {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    /*background-color: antiquewhite;*/
-    width: 60rem;
-    height: 30rem;
-    margin-top: 10rem;
-    border: 1px solid black;
-}
