+++ src/pages/chat/chat.scss
@@ -1,0 +1,10 @@
+.chat__container {
+  width: 60rem;
+  height: 30rem;
+  margin-top: 10rem;
+  border: 1px solid black;
+}
+
+.chat__container__content {
+  height: 100%;
+}
